# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Engage API",
    "version": "1.0.0",
    "description": "選挙活動管理プラットフォーム Engage のAPIドキュメント"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/api/leads-status": {
      "get": {
        "summary": "リードステータス一覧取得",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organizationId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "リードステータス作成",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードステータス作成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadsStatus"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "リードステータス更新",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードステータス更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadsStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{id}/activities/{activityId}": {
      "put": {
        "summary": "リードアクティビティ更新",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードアクティビティ更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadActivity"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "リードアクティビティ削除",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードアクティビティ削除",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{id}/activities": {
      "post": {
        "summary": "リードのアクティビティ作成",
        "description": "指定したリードに新しいアクティビティを作成します。",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "typeId": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成されたアクティビティ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadActivity"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "リードのアクティビティ一覧取得",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "leadId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードのアクティビティ一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadActivity"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{id}/groups": {
      "patch": {
        "summary": "リードのグループ一括更新",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "groupIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードのグループ一括更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lead"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LeadActivity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "leadId": { "type": "string" },
          "type": { "type": "string" },
          "description": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "typeId": { "type": "string" },
          "organizationId": { "type": "integer" }
        }
      },
...
      "Lead": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "organizationId": { "type": "integer" },
          "type": { "type": "string" },
          "name": { "type": "string" },
          "nameReading": { "type": "string" },
          "nickname": { "type": "string" },
          "district": { "type": "string" },
          "homePhone": { "type": "string" },
          "mobilePhone": { "type": "string" },
          "company": { "type": "string" },
          "position": { "type": "string" },
          "postalCode": { "type": "string" },
          "address": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "referrer": { "type": "string" },
          "evaluation": { "type": "integer" },
          "status": { "type": "string" },
          "statusId": { "type": "string" },
          "isPaid": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "LeadsStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "color": { "type": "string" },
          "organizationId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}